backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
web3==6.11.1
cryptography==41.0.4
openai>=1.0.0
anthropic>=0.3.11
gunicorn==21.2.0
